// server/src/payments/domain/enums/PaymentStatus.ts
export enum PaymentStatus {
  CREATED = 'CREATED',
  INITIATED = 'INITIATED',
  AUTHENTICATED = 'AUTHENTICATED',
  PENDING_VERIFICATION = 'PENDING_VERIFICATION',
  CAPTURED = 'CAPTURED',
  FAILED = 'FAILED',
  REFUNDED = 'REFUNDED',
  PARTIALLY_REFUNDED = 'PARTIALLY_REFUNDED',
  VOIDED = 'VOIDED'
}